Parallel Pipelines extension does not revert bash state correctly (#373)
* Add a logging prefix that shows whether we are in the daemon or pash.py


* Add a test that checks the behavior


* Solve the issue by reverting pash state


* Temporarily remove parallel pipelines since sometimes some tests get stuck

+++ pa.sh
@@ -42,5 +42,9 @@ then
   wait 2> /dev/null 1>&2 
 fi
 
-rm -rf "${PASH_TMP_PREFIX}"
+## Don't delete the temporary directory if we are debugging
+if [ "$PASH_DEBUG_LEVEL" -eq 0 ]; then
+  rm -rf "${PASH_TMP_PREFIX}"
+fi
+
 (exit $pash_exit_code)
